fftech-ai-website-audit-saas/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.24.0.post1
aiohttp==3.9.3
httpx==0.26.0
beautifulsoup4==4.12.2
lxml==5.1.0
pydantic==1.10.13
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.9
itsdangerous==2.1.2
PyJWT==2.8.0
APScheduler==3.10.4
reportlab==4.0.8
matplotlib==3.8.2
pandas==2.1.4
openpyxl==3.1.2
python-pptx==0.6.23
jinja2==3.1.2
